+++ src/system/post-office.ts
@@ -231,12 +231,12 @@ export class PostOffice {
      * 
      * @param event envelope
      * @param timeout value in milliseconds
-     * @returns a future promise of result or error
+     * @returns a future promise of result event or error
      */
     request(event: EventEnvelope, timeout = 60000): Promise<EventEnvelope> {
         this.touch(event);
         return self.request(event, timeout);
-    }  
+    } 
     
     /**
      * Make an asynchronous RPC call using "Event Over HTTP"
@@ -252,6 +252,20 @@ export class PostOffice {
         this.touch(event);
         return self.remoteRequest(event, endpoint, securityHeaders, rpc, timeout);   
     }
+
+    /**
+     * Make a fork-n-join RPC call to multiple services
+     * 
+     * @param events in a list
+     * @param timeout value in milliseconds
+     * @returns a future promise of a list of result events or error
+     */
+    parallelRequest(events: Array<EventEnvelope>, timeout = 60000): Promise<Array<EventEnvelope>> {
+        for (const event of events) {
+            this.touch(event);
+        }
+        return self.parallelRequest(events, timeout);
+    }   
 }
 
 export class Sender {
@@ -465,5 +479,39 @@ class PO {
                     reject(new AppException(status, e.message));
                 });
         });       
+    }
+
+    parallelRequest(events: Array<EventEnvelope>, timeout = 60000): Promise<Array<EventEnvelope>> {
+        return new Promise((resolve, reject) => {
+            // validate events
+            if (events.length == 0) {
+                reject(new AppException(400, 'Input must be not an empty list'));
+            }
+            for (const event of events) {
+                if (!(event instanceof EventEnvelope)) {
+                    reject(new AppException(400, 'Input must be a list of events'));
+                    return;
+                }
+            }
+            // process events
+            let normal = true;
+            const consolidated = new Array<EventEnvelope>();
+            for (const event of events) {
+                this.request(event, timeout)
+                    .then(result => {
+                        if (normal) {
+                            consolidated.push(result);
+                            if (consolidated.length == events.length) {
+                                resolve(consolidated);
+                            }
+                        }
+                    })
+                    .catch(e => {
+                        normal = false;
+                        const status = e instanceof AppException? e.getStatus() : 500;
+                        reject(new AppException(status, e.message));
+                    });
+            }
+        });
     }
 }